+++ Sources/BumpCommandLine/BumpCommand.swift
@@ -17,7 +17,7 @@ struct BumpCommand: ParsableCommand {
     @Option(name: .shortAndLong, help: "Increment mode to bump version or build number. Either 'major', 'minor', 'patch', or 'build'.")
     var mode: IncrementMode
 
-    @Option(name: .shortAndLong, help: "The path of .xcodeproj file or directory. Default value is the environment directory.")
+    @Option(name: .shortAndLong, help: "The path of .xcodeproj file or directory. Default value is the current directory.")
     var path: String?
 
     @Flag(name: .shortAndLong, help: "Show all the targets")
@@ -80,54 +80,6 @@ struct BumpCommand: ParsableCommand {
         }
 
         let directoryContents = try environment.fileManagerWrapper.contentsOfDirectory(
-            at: dirURL
-        )
-
-        guard let url = directoryContents.first(where: { $0.isXcodeProj }) else {
-            throw ValidationError("Needs exist a .xcodeproj file in this directory.")
-        }
-        return url.path
-    }
-}
-
-import FileManagerWrapper
-
-protocol XcodeProjFinderProtocol {
-    func findXcodeProj(path: String) throws -> String
-}
-
-struct XcodeProjFinder: XcodeProjFinderProtocol {
-    struct FindError: Error {
-        let message: String
-
-        init(_ message: String) {
-            self.message = message
-        }
-
-        var description: String {
-            message
-        }
-    }
-
-    let fileManagerWrapper: FileManagerWrapperProtocol
-
-    func findXcodeProj(path: String) throws -> String {
-        guard fileManagerWrapper.fileExists(atPath: path) else {
-            throw ValidationError("Needs exist a path of .xcodeproj file or directory.")
-        }
-
-        guard let dirURL = URL(string: path) else {
-            throw ValidationError("Wrong directory or path.")
-        }
-
-        if dirURL.isXcodeProj {
-            return dirURL.path
-        }
-        guard dirURL.pathExtension.isEmpty else {
-            throw ValidationError("The path must be .xcodeproj file or directory.")
-        }
-
-        let directoryContents = try fileManagerWrapper.contentsOfDirectory(
             at: dirURL
         )
